requirements.txt
django==3.2.8
django-environ==0.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt
drf-yasg==1.20.0
psycopg2-binary==2.9.1
requests==2.26.0

